BIOSTD-282 Lost properties in desirable subsections
Fixes bug where added subsections sometimes lose their properties.
This fixed that by repopulating `subSectionTypeMap` not only in the component initialization but also when adding a new subsection.
This change also includes the deletion of the corresponding entry in `subSectionTypeMap` when a subsection is deleted, to keep the model consistentcy